api/requirements.txt
flask==3.0.3
supabase==2.4.0
python-dotenv==1.0.1
requests==2.31.0
fastf1==3.3.0
